requirements.txt
pystray
watchdog
requests
Pillow
python-dotenv